Apply renames to DllImport and fix nuget package generation
