+++ src/SpectatorView.Native/.gitignore
@@ -44,7 +44,6 @@ artifacts/
 *_p.c
 *_i.h
 *.ilk
-*.meta
 *.obj
 *.pch
 *.pdb
