+++ src/SpectatorView.Native/CreateNuGetPackages.bat
@@ -1,0 +1,1 @@
+nuget pack SpectatorView.OpenCV.nuspec -Symbols
