--- src/SpectatorView.Native/PluginMetaFiles/.gitignore
@@ -1,5 +1,0 @@
-# dlls
-WSA/x64/*
-WSA/x86/*
-x64/*
-x86/*
