+++ src/SpectatorView.Native/SpectatorView.Native.sln
@@ -78,6 +78,7 @@ Global
 		{40A65FA5-46F2-4E86-92B9-CD61A6BED4DB}.Release|ARM.ActiveCfg = Release|ARM
 		{40A65FA5-46F2-4E86-92B9-CD61A6BED4DB}.Release|ARM.Build.0 = Release|ARM
 		{40A65FA5-46F2-4E86-92B9-CD61A6BED4DB}.Release|x64.ActiveCfg = Release|x64
+		{40A65FA5-46F2-4E86-92B9-CD61A6BED4DB}.Release|x64.Build.0 = Release|x64
 		{40A65FA5-46F2-4E86-92B9-CD61A6BED4DB}.Release|x86.ActiveCfg = Release|Win32
 		{40A65FA5-46F2-4E86-92B9-CD61A6BED4DB}.Release|x86.Build.0 = Release|Win32
 		{D8ECBA6D-7895-4ABA-879C-E18E18DFCB3C}.Debug|ARM.ActiveCfg = Debug|Win32
