+++ src/SpectatorView.Native/SpectatorView.OpenCV.nuspec
@@ -1,0 +1,35 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<package>
+  <metadata>
+    <id>Microsoft.SpectatorView.OpenCV</id>
+    <version>0.0.4</version>
+    <description>This is a test NuGet package definition for distributing the Spectator View binaries and their dependencies. Binaries will be output into the proper Unity Plugins special folder.</description>
+    <authors>Unknown</authors>
+    <licenseUrl>https://github.com/Microsoft/MixedRealityToolkit/blob/master/License.txt</licenseUrl>
+    <projectUrl>https://https://aka.ms/mrtk</projectUrl>
+    <iconUrl>https://www.nuget.org/Content/Images/packageDefaultIcon-50x50.png</iconUrl>
+    <releaseNotes>This is a test release</releaseNotes>
+    <copyright>Copyright 2019</copyright>
+    <tags>MRTK SpectatorView SpectatorViewDependencies</tags>
+  </metadata>
+
+  <files>
+    <file src="Release\SpectatorView.OpenCV.UWP\*.dll" target="lib\unity\Plugins\WSA\x86" />
+    <file src="Release\SpectatorView.OpenCV.UWP\*.pdb" target="lib\unity\Plugins\WSA\x86" />
+
+    <file src="x64\Release\SpectatorView.OpenCV.UWP\*.dll" target="lib\unity\Plugins\WSA\x64" />
+    <file src="x64\Release\SpectatorView.OpenCV.UWP\*.pdb" target="lib\unity\Plugins\WSA\x64" />
+
+    <file src="x64\Release\SpectatorView.OpenCV.dll" target="lib\unity\Plugins\x64" />
+    <file src="x64\Release\SpectatorView.OpenCV.pdb" target="lib\unity\Plugins\x64" />
+    <file src="x64\Release\opencv_*.dll" target="lib\unity\Plugins\x64" />
+    <file src="x64\Release\jpeg62.dll" target="lib\unity\Plugins\x64" />
+    <file src="x64\Release\libpng16.dll" target="lib\unity\Plugins\x64" />
+    <file src="x64\Release\lzma.dll" target="lib\unity\Plugins\x64" />
+    <file src="x64\Release\zlib1.dll" target="lib\unity\Plugins\x64" />
+
+    <file src="UnityMetaFiles\x64\*.meta" target="lib\unity\Plugins\x64" />
+    <file src="UnityMetaFiles\WSA\x86\*.meta" target="lib\unity\Plugins\WSA\x86" />
+    <file src="UnityMetaFiles\WSA\x64\*.meta" target="lib\unity\Plugins\WSA\x64" />
+  </files>
+</package>
