+++ src/SpectatorView.Native/UnityMetaFiles/.gitignore
@@ -1,0 +1,3 @@
+# Re-allow meta files in these directories
+!x86/
+!x64/
