+++ src/SpectatorView.Unity/Assets/SpectatorView/Scripts/MarkerDetection/ArUco/HoloLens/SpectatorViewOpenCVInterface.cs.meta
